requirements.txt
attrs==20.2.0
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.2
github3.py==1.3.0
idna==2.9
imageio==2.8.0
importlib-metadata==2.0.0
iniconfig==1.0.1
jwcrypto==0.7
kiwisolver==1.2.0
matplotlib==3.2.2
networkx==2.4
numpy==1.18.5
opencv-python==4.2.0.34
packaging==20.4
pandas==1.1.1
Pillow==7.1.2
pip-tools==5.2.1
pluggy==0.13.1
psycopg2==2.8.6
py==1.9.0
pycparser==2.20
pyparsing==2.4.7
pytesseract==0.3.4
pytest==6.1.1
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
requests==2.24.0
scikit-image==0.17.2
scipy==1.1.0
six==1.15.0
soupsieve==2.0.1
tifffile==2020.6.3
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.9
zipp==3.3.0
